continue where left download

+++ utils/download_dataset.py
@@ -26,6 +26,7 @@ def main(args):
     basedir = data["basedir"]
     relpaths = data["relpaths"]
     if args.num_demos is not None:
+        # np.random.randint(low=0, high=len(relpaths), size=args.demos)
         relpaths = relpaths[:args.num_demos]
 
     if not os.path.exists(args.output_dir):
@@ -37,21 +38,25 @@ def main(args):
         outpath = os.path.join(args.output_dir, filename)
         percent_done = 100 * i / len(relpaths)
         print(f"[{percent_done:.0f}%] Downloading {outpath}")
-        try:
-            urllib.request.urlretrieve(url, outpath)
-        except Exception as e:
-            print(f"\tError downloading {url}: {e}. Moving on")
-            continue
+        if not os.path.exists(outpath):
+            # print("\t Already Exists")
+            try:
+                urllib.request.urlretrieve(url, outpath)
+            except Exception as e:
+                print(f"\tError downloading {url}: {e}. Moving on")
+                continue
 
         # Also download corresponding .jsonl file
         jsonl_url = url.replace(".mp4", ".jsonl")
         jsonl_filename = filename.replace(".mp4", ".jsonl")
         jsonl_outpath = os.path.join(args.output_dir, jsonl_filename)
-        try:
-            urllib.request.urlretrieve(jsonl_url, jsonl_outpath)
-        except Exception as e:
-            print(f"\tError downloading {jsonl_url}: {e}. Cleaning up mp4")
-            os.remove(outpath)
+        if not os.path.exists(jsonl_outpath):
+            # print("\t Already Exists")
+            try:
+                urllib.request.urlretrieve(jsonl_url, jsonl_outpath)
+            except Exception as e:
+                print(f"\tError downloading {jsonl_url}: {e}. Cleaning up mp4")
+                os.remove(outpath)
 
 
 if __name__ == "__main__":
